requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
distro==1.3.0
Flask==1.0.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.14.3
pandas==0.23.0
python-dateutil==2.7.3
pytz==2018.4
requests==2.18.4
six==1.11.0
tabula-py==1.2.0
urllib3==1.22
Werkzeug==0.14.1
gunicorn==19.7.1
gevent==1.2.2
